vertical list with next 5 days and horizontal with 3 hour forecast by day

+++ app/src/main/java/pt/fabiomatos/weatherapp/models/Daily.kt
@@ -40,7 +40,7 @@ data class Daily(
     @SerializedName("clouds")
     var clouds: Int? = null,
     @SerializedName("pop")
-    var pop: Int? = null,
+    var pop: Float? = null,
     @SerializedName("uvi")
     var uvi: Double? = null
