+++ app/src/main/java/pt/fabiomatos/weatherapp/models/ListDay.kt
@@ -12,7 +12,7 @@ data class ListDay(
     @SerializedName("clouds") var clouds: Clouds? = Clouds(),
     @SerializedName("wind") var wind: Wind? = Wind(),
     @SerializedName("visibility") var visibility: Int? = null,
-    @SerializedName("pop") var pop: Int? = null,
+    @SerializedName("pop") var pop: Float? = null,
     @SerializedName("sys") var sys: Sys? = Sys(),
     @SerializedName("dt_txt") var dtTxt: String? = null
